feat: search and reset data
