+++ src/components/button.tsx
@@ -1,0 +1,22 @@
+import React from 'react';
+
+interface ButtonProps {
+  label: string;
+  onClick: () => void;
+  disabled?: boolean;
+  className?: string;
+}
+
+const Button: React.FC<ButtonProps> = ({ label, onClick, disabled, className }) => {
+  return (
+    <button
+      onClick={onClick}
+      disabled={disabled}
+      className={`bg-blue-500 hover:bg-blue-700 text-white font-bold py-2 px-4 rounded cursor-pointer ${className}`}
+    >
+      {label}
+    </button>
+  );
+};
+
+export default Button;
